+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/ext/LinkingObjectsExt.kt
@@ -1,10 +1,10 @@
 package io.realm.kotlin.ext
 
-import io.realm.kotlin.internal.RealmLinkingObjectsImpl
-import io.realm.kotlin.types.RealmLinkingObjects
+import io.realm.kotlin.internal.RealmLinkingObjectsDelegateImpl
+import io.realm.kotlin.types.RealmLinkingObjectsDelegate
 import io.realm.kotlin.types.RealmObject
 import kotlin.reflect.KProperty1
 
 // TODO document
-public fun <T : RealmObject> linkingObjects(targetProperty: KProperty1<T, *>): RealmLinkingObjects<T> =
-    RealmLinkingObjectsImpl()
+public fun <T : RealmObject> linkingObjects(targetProperty: KProperty1<T, *>): RealmLinkingObjectsDelegate<T> =
+    RealmLinkingObjectsDelegateImpl()