+++ packages/test-base/src/androidTest/kotlin/io/realm/kotlin/test/shared/LinkingObjectsTests.kt
@@ -20,25 +20,25 @@ package io.realm.kotlin.test.shared
 
 import io.realm.kotlin.Realm
 import io.realm.kotlin.RealmConfiguration
-import io.realm.kotlin.entities.Sample
 import io.realm.kotlin.ext.isManaged
 import io.realm.kotlin.ext.linkingObjects
 import io.realm.kotlin.ext.query
 import io.realm.kotlin.ext.realmListOf
 import io.realm.kotlin.ext.realmSetOf
-import io.realm.kotlin.test.platform.PlatformUtils
-import io.realm.kotlin.types.RealmList
-import io.realm.kotlin.types.RealmObject
 import io.realm.kotlin.internal.asDynamicRealm
 import io.realm.kotlin.notifications.InitialResults
 import io.realm.kotlin.notifications.ResultsChange
 import io.realm.kotlin.notifications.UpdatedResults
 import io.realm.kotlin.query.RealmResults
+import io.realm.kotlin.test.platform.PlatformUtils
+import io.realm.kotlin.types.RealmList
+import io.realm.kotlin.types.RealmObject
 import io.realm.kotlin.types.RealmSet
 import io.realm.kotlin.types.RealmUUID
 import io.realm.kotlin.types.annotations.Ignore
 import kotlinx.coroutines.async
 import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.flow.onCompletion
 import kotlinx.coroutines.runBlocking
 import kotlin.test.BeforeTest
 import kotlin.test.Test
@@ -48,7 +48,9 @@ import kotlin.test.assertIs
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
 
-class Parent : RealmObject {
+class Parent(var id: Int) : RealmObject {
+    constructor() : this(0)
+
     var child: Child? = null
     var childList: RealmList<Child> = realmListOf()
     var childSet: RealmSet<Child> = realmSetOf()
@@ -63,8 +65,8 @@ class Child : RealmObject {
 class Recursive : RealmObject {
     var name: RealmUUID = RealmUUID.random()
 
-    var recursive: Recursive? = null
-    val references by linkingObjects(Recursive::recursive)
+    var recursiveField: Recursive? = null
+    val references by linkingObjects(Recursive::recursiveField)
 }
 
 class MissingSourceProperty : RealmObject {
@@ -116,7 +118,6 @@ class LinkingObjectsTests {
         }
     }
 
-
     // TODO not part of the final test suite, its is here to validate that the users can write their own code
     private val Child.parentsOrNull: RealmResults<Parent>?
         get() = if (isManaged()) parents else null
@@ -134,7 +135,7 @@ class LinkingObjectsTests {
                 val child = this.copyToRealm(Child())
 
                 val parents = Array(5) {
-                    this.copyToRealm(Parent())
+                    this.copyToRealm(Parent(it))
                 }
 
                 child.assertParents(0)
@@ -157,7 +158,7 @@ class LinkingObjectsTests {
                 val child = this.copyToRealm(Child())
 
                 val parents = Array(5) {
-                    this.copyToRealm(Parent())
+                    this.copyToRealm(Parent(it))
                 }
 
                 child.assertParents(0)
@@ -182,7 +183,7 @@ class LinkingObjectsTests {
         runBlocking {
             val recursive = realm.write {
                 val recursive = this.copyToRealm(Recursive())
-                recursive.recursive = recursive
+                recursive.recursiveField = recursive
                 recursive
             }
             assertEquals(1, recursive.references.size)
@@ -197,11 +198,14 @@ class LinkingObjectsTests {
             realm.write {
                 val child = this.copyToRealm(Child())
 
-                this.copyToRealm(Parent().apply {
-                    this.child = child
-                    this.childSet.add(child)
-                    this.childList.add(child)
-                })
+                this.copyToRealm(
+                    Parent().apply
+                    {
+                        this.child = child
+                        this.childSet.add(child)
+                        this.childList.add(child)
+                    }
+                )
             }
 
             realm.asDynamicRealm().query(Child::class.simpleName!!).first().find()!!
@@ -260,99 +264,107 @@ class LinkingObjectsTests {
     }
 
     @Test
-    fun migration() {
+    fun migration(): Unit = TODO()
 
-    }
-
-    // Smoke tests
+    // Notification smoke tests
     @Test
-    fun notifications_objectLink() {
+    fun notifications_source_object() {
         runBlocking {
             val child = realm.write {
                 copyToRealm(Child())
             }
 
-            val c = Channel<ResultsChange<Parent>>(1)
+            val c = Channel<ResultsChange<Parent>?>(1)
             val observer = async {
                 realm.query<Child>()
                     .first()
                     .find()!!
                     .parents
                     .asFlow()
+                    .onCompletion {
+                        c.trySend(null)
+                    }
                     .collect {
                         c.trySend(it)
                     }
             }
 
-            c.receive().let { resultsChange ->
+            c.receive()!!.let { resultsChange ->
                 assertIs<InitialResults<*>>(resultsChange)
                 assertTrue(resultsChange.list.isEmpty())
             }
 
             realm.write {
                 repeat(5) {
-                    copyToRealm(Parent()).apply {
+                    copyToRealm(Parent(it)).apply {
                         this.child = findLatest(child)
                     }
                 }
             }
 
-            c.receive().let { resultsChange ->
+            c.receive()!!.let { resultsChange ->
                 assertIs<UpdatedResults<*>>(resultsChange)
                 assertEquals(5, resultsChange.list.size)
             }
 
+            // Validate the flow completes once the target object gets deleted
             realm.write {
                 delete(findLatest(child)!!)
             }
 
-            c.receive().let { resultsChange ->
-                assertIs<UpdatedResults<*>>(resultsChange)
-                assertEquals(0, resultsChange.list.size)
-            }
+            assertNull(c.receive())
 
             observer.cancel()
             c.close()
         }
     }
 
-
     @Test
-    fun notifications_lists() {
+    fun notifications_source_list() {
         runBlocking {
             val child = realm.write {
                 copyToRealm(Child())
             }
 
-            val c = Channel<ResultsChange<Parent>>(1)
+            val c = Channel<ResultsChange<Parent>?>(1)
             val observer = async {
                 realm.query<Child>()
                     .first()
                     .find()!!
                     .parentsByList
                     .asFlow()
+                    .onCompletion {
+                        c.trySend(null)
+                    }
                     .collect {
                         c.trySend(it)
                     }
             }
 
-            c.receive().let { resultsChange ->
+            c.receive()!!.let { resultsChange ->
                 assertIs<InitialResults<*>>(resultsChange)
                 assertTrue(resultsChange.list.isEmpty())
             }
 
             realm.write {
                 repeat(5) {
-                    copyToRealm(Parent()).apply {
+                    copyToRealm(Parent(it)).apply {
                         this.childList.add(findLatest(child)!!)
                     }
                 }
             }
 
-            c.receive().let { resultsChange ->
+            c.receive()!!.let { resultsChange ->
                 assertIs<UpdatedResults<*>>(resultsChange)
                 assertEquals(5, resultsChange.list.size)
             }
+
+            // Validate the flow completes once the target object gets deleted
+            realm.write {
+                delete(findLatest(child)!!)
+            }
+
+            assertNull(c.receive())
 
             observer.cancel()
             c.close()
@@ -360,49 +372,82 @@ class LinkingObjectsTests {
     }
 
     @Test
-    fun notifications_sets() {
+    fun notifications_source_set() {
         runBlocking {
             val child = realm.write {
                 copyToRealm(Child())
             }
 
-            val c = Channel<ResultsChange<Parent>>(1)
+            val c = Channel<ResultsChange<Parent>?>(1)
             val observer = async {
                 realm.query<Child>()
                     .first()
                     .find()!!
                     .parentsBySet
                     .asFlow()
+                    .onCompletion {
+                        c.trySend(null)
+                    }
                     .collect {
                         c.trySend(it)
                     }
             }
 
-            c.receive().let { resultsChange ->
+            c.receive()!!.let { resultsChange ->
                 assertIs<InitialResults<*>>(resultsChange)
                 assertTrue(resultsChange.list.isEmpty())
             }
 
             realm.write {
                 repeat(5) {
-                    copyToRealm(Parent()).apply {
+                    copyToRealm(Parent(it)).apply {
                         this.childSet.add(findLatest(child)!!)
                     }
                 }
             }
 
-            c.receive().let { resultsChange ->
+            c.receive()!!.let { resultsChange ->
                 assertIs<UpdatedResults<*>>(resultsChange)
                 assertEquals(5, resultsChange.list.size)
             }
+
+            // Validate the flow completes once the target object gets deleted
+            realm.write {
+                delete(findLatest(child)!!)
+            }
+
+            assertNull(c.receive())
 
             observer.cancel()
             c.close()
         }
     }
 
+    // queries smoke tests
     @Test
     fun queries() {
-        // Test that we can query the results
+        realm.writeBlocking {
+            val child = this.copyToRealm(Child())
+
+            val parents = Array(5) {
+                this.copyToRealm(Parent(it))
+            }
+
+            child.assertParents(0)
+
+            parents.forEach { parent ->
+                parent.child = child
+                parent.childList.add(child)
+                parent.childSet.add(child)
+            }
+
+            child.assertParents(parents.size)
+        }
+
+        realm.query<Child>().first().find()!!.let { child ->
+            assertEquals(2, child.parents.query("id > 2").find().size)
+            assertEquals(2, child.parentsByList.query("id > 2").find().size)
+            assertEquals(2, child.parentsBySet.query("id > 2").find().size)
+        }
     }
 }